jgt-rbac-case/jgt-tasksys/src/main/java/org/hepx/tasksys/service/OrganizationService.java
package org.hepx.tasksys.service;

import org.hepx.tasksys.entity.Organization;

import java.util.List;

public interface OrganizationService {


    public Organization createOrganization(Organization organization);

    public Organization updateOrganization(Organization organization);

    public void deleteOrganization(Long organizationId);

    Organization findOne(Long organizationId);

    List<Organization> findAll();

    List<Organization> findAllWithExclude(Organization excludeOraganization);

    void move(Organization source, Organization target);
}
